[feat: ValidationInput 기능 수정]
