+++ src/components/ValidationInput/index.html
@@ -8,58 +8,17 @@
     <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
     <link rel="icon" sizes="192x192" href="/android-icon.png" />
     <link rel="stylesheet" href="/src/styles/tailwind.css" />
+    <script
+      type="module"
+      src="/src/pages/login/test/ValidationInput.js"
+    ></script>
     <title>유효성검사 인풋</title>
   </head>
   <body class="min-h-screen">
     <div class="relative min-w-screen max-w-screen w-full mx-auto">
-      <div class="p-3">
-        <label for="label1" class="text-label-sm">레이블명1</label>
-        <div class="flex flex-col w-full">
-          <input
-            type="text"
-            id="label1"
-            class="input p-2 text-paragraph-sm font-normal w-full outline-none border border-contents-content-tertiary rounded"
-            placeholder="플레이스홀더 삽입"
-            spellcheck="false"
-          />
-          <div class="flex justify-between">
-            <span
-              class="error-msg text-red-500 text-paragraph-sm font-normal opacity-0"
-            ></span>
-            <span
-              aria-live="polite"
-              class="letter-count text-contents-content-tertiary text-paragraph-sm font-normal"
-            ></span>
-          </div>
-        </div>
-      </div>
-      <div class="p-3">
-        <label for="label2" class="text-label-sm">레이블명2</label>
-        <div class="flex flex-col w-full">
-          <input
-            type="text"
-            id="label2"
-            class="input p-2 text-paragraph-sm font-normal w-full outline-none border border-contents-content-tertiary rounded"
-            placeholder="플레이스홀더 삽입"
-            spellcheck="false"
-          />
-          <div class="flex justify-between">
-            <span
-              class="error-msg text-red-500 text-paragraph-sm font-normal opacity-0"
-            ></span>
-            <span
-              aria-live="polite"
-              class="letter-count text-contents-content-tertiary text-paragraph-sm font-normal"
-            ></span>
-          </div>
-        </div>
-      </div>
+      <div class="valid-input-box"></div>
+      <div class="valid-input-box"></div>
+      <div class="valid-input-box"></div>
     </div>
-
-    <script
-      type="module"
-      src="/src/components/ValidationInput/ValidationInput.js
-    "
-    ></script>
   </body>
 </html>
