+++ src/components/ValidationInput/index2.html
@@ -1,0 +1,65 @@
+<!doctype html>
+<html lang="ko">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="shortcut icon" sizes="48x48" href="/favicon.ico" />
+    <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
+    <link rel="icon" sizes="192x192" href="/android-icon.png" />
+    <link rel="stylesheet" href="/src/styles/tailwind.css" />
+    <title>유효성검사 인풋</title>
+  </head>
+  <body class="min-h-screen">
+    <div class="relative min-w-screen max-w-screen w-full mx-auto">
+      <div id="label1" class="valid-input-box p-3">
+        <label for="label1" class="text-label-sm">레이블명1</label>
+        <div class="flex flex-col w-full">
+          <input
+            type="text"
+            id="label1"
+            class="input p-2 text-paragraph-sm text w-full outline-none border border-contents-content-tertiary rounded"
+            placeholder="플레이스홀더 삽입"
+            spellcheck="false"
+          />
+          <div class="flex justify-between">
+            <span
+              class="error-msg text-red-500 text-paragraph-sm font-normal opacity-0"
+            ></span>
+            <span
+              aria-live="polite"
+              class="letter-count text-contents-content-tertiary text-paragraph-sm font-normal"
+            ></span>
+          </div>
+        </div>
+      </div>
+      <div class="p-3">
+        <label for="label2" class="text-label-sm">레이블명2</label>
+        <div class="flex flex-col w-full">
+          <input
+            type="text"
+            id="label2"
+            class="input p-2 text-paragraph-sm font-normal w-full outline-none border border-contents-content-tertiary rounded"
+            placeholder="플레이스홀더 삽입"
+            spellcheck="false"
+          />
+          <div class="flex justify-between">
+            <span
+              class="error-msg text-red-500 text-paragraph-sm font-normal opacity-0"
+            ></span>
+            <span
+              aria-live="polite"
+              class="letter-count text-contents-content-tertiary text-paragraph-sm font-normal"
+            ></span>
+          </div>
+        </div>
+      </div>
+    </div>
+
+    <script
+      type="module"
+      src="/src/components/ValidationInput/ValidationInput.js
+    "
+    ></script>
+  </body>
+</html>
